CODSOFT TASK 2
This project focuses on building a machine learning model that predicts the rating of a movie based on features such as genre, director, and cast. By analyzing historical movie data, the goal is to understand the key factors that influence movie ratings and accurately estimate the ratings for new or upcoming films.